moved the history to wallet status section
