+++ app/renderer/assets/locales/en-US/wallet.json
@@ -30,6 +30,7 @@
   "wallet_status_resources_usage": "Usage",
   "wallet_status_tab_token_balances": "Token Balances",
   "wallet_status_tab_staked": "Staked EOS",
+  "wallet_status_tab_account_history": "Account History",
   "wallet_status_tab_account_data": "Account Data",
   "wallet_panel_form_create_wallet_header": "Import your Account",
   "wallet_panel_form_create_wallet_body": "To use eos-voter to vote, fill in your account name and private key.",
